Rename Footer -> MageFooter
Kind of a hacky name, but multi-word names are now enforced.
