+++ src/components/MageFooter.vue
